(SERVER-1412) Allow build scripts to be included in staging tarball
This commit adds support for creating an `ext/build-scripts/` dir
in a project's jar.  If found, ezbake will copy all of the files
found in this directory into the staging directory, so that they
will be included in the tarball that is sent up to the packaging
system.

These files will *not* be included in the final packages, but they
can be referenced during the package build, e.g. during the `install`
phase, in order to reduce duplication of logic in the ezbake config
files.