(EZ-103) Address incorrect rundir perms
Prior to this commit, we had been creating the rundir, but sometimes
with permissions which prevented the service from being able to create
its own pidfile.  In these cases, the Java process could be launched
successfully and the service start command could still return success.
Any subsequent attempts to ask the service framework for status on the
service, however, would return failure (not started).

This commit should ensure that the rundir is always created with
permissions which allow the service to create its own pidfile.  This
commit also adds some error handling to return failure at service
startup if the pidfile unexpectedly could not be written for some
reason.