(EZ-102) Update EZBake to install SysV + SystemD
This commit updates EZBake to start installing sysv files on newer
systemd based debian-ish systems. The idea being that users should be
able to switch between the older and newer service management tools.
This is a behavior that has already been established in other packages
that are targeted at newer releases.

This commit also updates EZBake to provide native SystemD files for
Debian 8 Jessie. This had not previously been done because of concerns
over saving state on upgrade from SysV to SystemD. This is no longer a
concern.