(SERVER-1670) Stop autogenerating dependencies for RPM packages
Currently the generated RPM packages will automatically generate
a set of requirements based on the contents of the package. This
has become an issue with puppet-server 2.7.0, as adding gems
to the package inadvertently added a dependency on system ruby.

This adds 'AutoReq: 0' to the RPM specs to stop generating automatic
dependency information, and adds 'bash' as the only dependency
that was being automatically added that we still want.

This also adds 'AutoProv: 0' to stop generating a package 'provides'.
This should be unnecessary, but harmless and may save us some pain in
the future if we wind up shipping any shared libs in a package somewhere.