(SERVER-1723) Use /dev/urandom for foreground subcommand
This commit adds -Djava.security.egd=/dev/urandom as an argument to the
Java command line for the foreground subcommand.  This argument was
already being used in the service start subcommand.  Performance for some
operations like SSL can be significantly faster with this option in
place so it seems like having this benefit when troubleshooting with the
foreground subcommand would be a good thing.