(SERVER-1763) Add ca-certificates as build dependency
Now that puppetserver has started bundling gems into the ezbake build,
java's lack of truststore anchors in our build has become a problem. In
jruby, simply requiring openssl will fail if there are no trust anchors.
This is the case on sles12. To avoid this problem, this commit adds the
correct ca-certificates package for el and for sles 12.