(maint) Resolve snapshot dep versions.
During staging, replace all SNAPSHOT dependency versions with the
most recently deployed version. This makes the builds transparent since
we'll know exactly which maven artifacts are in it, and improves
reproducibility since there must be a deployed snapshot version (locally
deployed snapshots are not permitted). If no deployed version of
a snapshotted dependency is found, EZBake throws an error.