(EZ-108) Allow staging w/undeployed snapshots
Add an options map to the functions in `ezbake.dependency-utils` that
expand snapshot versions to provide the option to use undeployed
snapshot artifacts by setting the :reproducible? option to `false`.

In the stage action, check whether the
`EZBAKE_ALLOW_UNREPRODUCIBLE_BUILDS` environment variable is set, and if
so set the :reproducible option to `false` to allow staging with
undeployed SNAPSHOT versions.