(RE-8726) Allow nonfinal repo definition
This commit allows ezbake to take advantage of a new feature in the
packaging repo. We can switch between a final and non-final repo
definition. This commit allows ezbake to set a value for that non-final
repo and for that setting to propagate through properly to the
automation in the packaging repo.

This means we can ship development builds with the same automation we
use to ship final builds. But the builds will be kept separate based on
if it's a final or non-final build (determined by the presence of
SNAPSHOT in the version name).

This is all in preparation for the change over to the puppet5 repos.
Before we have all the finalized puppet5 packages, we want to make
beta/development/nightly builds available on apt.puppetlabs.com,
yum.puppetlabs.com, and downloads.puppetlabs.com under a separate repo
than the main packages.