(EZ-109) Fix duplicate text in ezbake.manifest
This commit fixes an issue where some duplicate text would show up in
ezbake.manifest if additional uberjars were specified in the ezbake section of
a project.clj file

This was because of a combination of mustache being mustache, and because
not-empty returns its argument if it is not empty, rather than returning true.
This causes the mustache template to act unexpectedly.

(not (empty? ...)) always returns a boolean and fixes the issue