(EZ-110) Add immediate project jar into additional uberjar
Previously, the additional uberjar functionality appended the immediate
project jar as a dependency of itself in order for its content to be
generated into the uberjar.  This logic, however, did not work when the
project had any dependency overrides in the uberjar profile.  In that
case, none of the content from the immediate jar ended up in the uberjar
and the ezbake.manifest dependency info errantly listed info from the
main project's dependencies which did not end up in the uberjar.

In this commit, the immediate project's jar is appended directly into
the process of creating the uberjar - rather than as a self-referential
dependency.  This allows for the uberjar to properly retain content from
the main project jar even in the case of uberjar profile overrides.  The
main project's own artifact info is also added to the dependency listing
info directly, avoiding the appearance of errant transitive dependency
info in the output in cases where dependency exclusions are used.