(EZ-110) Retain qualifying coordinate info when snapshot version expanded
Previously, work was done to expand SNAPSHOT versions to the version of
the actual artifact previously deployed to a repository.  In the process
of expansion, any qualifying info on the coordinate, e.g., any
exclusions, was lost.  The resulting uberjar which was built would still
honor the qualifying info, however, the ezbake.manifest would errantly
list info that should have been excluded per the qualifying coordinate
info.

In this commit, the SNAPSHOT expansion logic only updates the dependency
version field, leaving all other trailing coordinate info in place.
This enables the ezbake.manifest to accurately reflect the content
generated into the uberjar.