(RE-8457) Update ezbake to ship projects to puppet5
This commit switches ezbake to use the 1.0.x branch of
puppetlabs/packaging, which contains all the updates we need to ship
packages to the new puppet5 and puppet5-nightly repos. This commit also
adds in the new variable `gpg_nonfinal_key`, which allows us to sign
nonfinal/development/beta/nightly builds with a separate key.

I'm also allowing the repo links to default to 'puppet' and
'puppet-nightly' respectively. This will allow users to overwrite these
values or set them to nil if they decide. But it'll make things a lot
easier for us in the future.