(SERVER-2068) Be more discerning finding PIDs
Previously, the start/stop/reload scripts would use a `pgrep` to search
for running processes, to determine if the service was still running.
That `pgrep` would do a simple search for uberjar name.

Turns out, that's not specific enough. In the case of puppetserver, a
user invocation of `puppetserver gem list` will put a process in the
process list that the naive `pgrep` would finger as the service pid in
the event puppetserver wasn't actually running. This could trick a
service manager into thinking puppetserver was running when in fact it
could be stopped for hours.

This commit doesn't change the basic service detection logic, but it
improves the `pgrep` to be more discerning, and search not only for the
uberjar name but also for the main namespace.