(CPR-515) Allow separate tasks for install vs upgrade
Currently, the postinstall tasks from the packaged project
are executed unconditionally, both on install as well as
upgrades. This can cause issues, as in the case with puppetserver
potentially overwriting config values if the user has changed them
from the package defaults.

This adds an addtional set of tasks that can be executed in the
package post-install script, that are only executed on initial
package install. This leaves the current behavior untouched, so
packages depending on the current behavior can be unmodified.