(EZ-124) Use consistent timestamps
We use timestamps for two separate parts of the `stage` action. Since
the timestamps only record the time to the minute, this only causes
issues when the timing is just right, but when it does it leads to
a mismatch in package version and the directory the packages are shipped
to.

This commit updates the stage action to generate the timestamp once and
pass it through to the functions that ultimately end up setting the
release field in ezbake.rb and tagging the git repo so we can be sure
we're using the same timestamp for both.