(RE-11573) Add fpm to Gemfile
This commit adds the 'fpm' gem to the Gemfile that gets packaged with our
Clojure projects. Now that we are using packaging as a gem, and thus using
bundler, we're calling all our tasks with 'bundle exec'. Even though fpm is
installed on our build container, the executable wasn't being found because fpm
was not included in the bundle.