(RE-11573) Hard-code rubylibdir in install script
This commit updates the `rubylibdir` variable used in the `install.sh` script
to a hard-coded value (`/opt/puppetlabs/puppet/lib/ruby/vendor_ruby`).
Previously, this variable was set using a call to puppet's ruby and checking
the `RbConfig::CONFIG['vendordir']`. Now that we are using packaging as a gem,
and thus bundler, we encountered errors in this script because puppet's ruby
does not have access to our bundle. This value doesn't contain any version info
and we don't expect it to change very frequently, so hard-coding this value
seemed like the easiest way to avoid these bundler issues.