(EZ-127) Add docker-build to build an image from a clojure repo
This adds a new `lein ezbake docker-build` command which can be used to
generate a standardized docker image from a trapperkeeper project.

The resulting image has the requisite java invocation set as its
entrypoint, referring a bootstrap.cfg and conf.d both located at
/config, and setting the classpath set to the project uberjar and any
additional jars, which are placed into /src. The base image defaults to
openjdk:8-jre-alpine, but can be overridden in the project.clj. Ports
to be exposed can be included in project.clj, as well as the name to tag
the image with, which defaults to the project name.