(EZ-128) Add support for specifying numeric uid/gid in rpm packages
For RPM packaging, we have a reserved uid/gid for the 'puppet' user. To
enable us to use the reserved uid/gid for the puppet user, ezbake needs
to support an option for that. The new option is `numeric-uid-gid`,
which can be set in the `:lein-ezbake` section of your project.clj.

This option is a noop for deb packaging. According to the debian policy
guide
(https://www.debian.org/doc/debian-policy/ch-opersys.html#uid-and-gid-classes)
reserved uids and gids need to be set in the `base-passwd` package.