(RE-12082) Change `lpasswd` back to `passwd`
As part of PE-24606, `adduser` commands were replaced with their `libuser`
counterparts (e.g. `useradd` -> `luseradd`). The database `passwd` was
incorrectly replaced with `lpasswd` as part of this change, which resulted in
installation and service failures for puppetserver, since there is no `lpasswd`
database. This commit puts `passwd` back in place of `lpasswd`.