(RE-12082) Use adduser commands on debian
This commit reverts some of the work done for PE-24606. We previously switched
from using `adduser` commands to using their `libuser` equivalents (e.g.
`useradd` -> `luseradd`), but there is a bug in `libuser1`, a dependency of
`libuser`, on debian 8 (jessie) and ubuntu 16.04 (xenial) that prevents these
commands from being used. This commit puts back the `adduser` commands for all
debian platforms. In theory, we can just revert this whole commit when we stop
supporting jessie and xenial and things should Just Work (tm).
For more info on the bug, see
https://bugs.launchpad.net/ubuntu/+source/system-config-samba/+bug/1718202.