(RE-12082) Revert libuser changes
This commit reverts the changes made to address PE-24606. We encountered
installation failure when bumping puppetserver to a version of ezbake with
these changes. There were several issues with this work, the largest of which
being the fact that there is no libuser package for sles and there is a bug in
libuser1, a dependency of libuser, for debian 8 (jessie) and ubuntu 16.04
(xenial) (https://bugs.launchpad.net/ubuntu/+source/system-config-samba/+bug/1718202).
Working around these issues (i.e. only using libuser commands on rhel systems)
is possible, but there are several unused code paths in this repo that are
making this hard to debug. The plan is to remove those paths before retrying
this work.

Revert "add libuser dependency in the correct spot"
This reverts commit c4e988049b128e8bd4f44f59e0cb8969bef3f028.

Revert "add libuser and update commands"
This reverts commit 16c5334a51cb9f7021e74e712814dd330668be63.

Revert "(PE-24606) use luseradd instead of useradd for ldap reboot issue"
This reverts commit 69727d1855f86e00f154ca6e1fab895d7e94be87.