(maint) Fix make-template-map test
The test was passing the wrong number of arguments for the current
definition, and it was also expecting an exception that is no longer
thrown (cf. 2a876d064c1acfc1ae08a8668a4fc262bb2fad7d).

Update the arguments and rework the test to just verify the result for
a trivial project.