(PDB-4373) ezbake.rb: insert complete, escaped ruby literals
Previously the template values were split, with some of the syntax
being in the template and some generated by clojure, and the values
were not escaped, leaving them susceptible to trouble if, for example
a string value included a double
quote (cf. https://www.xkcd.com/327/).

Add as-ruby-literal, which is intended to generate complete, correct
ruby literal values for the types it understands, and use it for all
of the expansion values.  This should allow it to correctly handle
values like this:

  debian: { preinst: ["do-something \"$some_var\""] ...