(EZ-35) Make use of project build-scripts
This commit adds a method to copy build-scripts from the current project to the
staging dir. Previously, we only copied build-scripts from direct dependencies,
which resulted in the FOSS puppetserver build-scripts to be used instead of the
pe-puppet-server-extensions build-scripts when building from pe-pse instead of
from pe-puppetserver (PE-28810).

It appears that puppetserver gets around this by declaring a dependency on
itself in the :ezbake profile.