(EZ-140) Allow excluding projects from ezbake resource search
Previously, ezbake would search all of the dependencies that it compiled
into the uberjar for ezbake config resources (e.g. CLI scripts and conf
files). However, we have a usecase now where one ezbake project is
wrapping another standalone ezbake project, so we want to exclude the
config from the wrapped project (puppetserver) and just use the fixtures
from the wrapping project (pe-puppetserver).

This commit adds a value, `excluded-ezbake-deps` to the `lein-ezbake`
config section, which contains a list of project symbols. These projects
will not be searched for packaging resources, but they will still be
included for compiling.