(EZ-142) Fix OnOutOfMemoryError flag used for older Javas
This commit makes a small correction to the OnOutOfMemoryError flag
used when starting JVMs if the Java version is older than 11.
Prior to this commit, the nested quoting of the argument for the
OnOutOfMemoryError flag resulted in Java attempting to execute
"kill -9 %p" as a single binary name instead of executing "kill"
and passing "-9" and "%p" as arguments.