(PDB-5622) Select Java 11+ on Platform 8
Does not change PE package builds.

Add a puppet-platform-version argument to EZBake defaulting to 7 (which
maintains the current dependency list) when building for Puppet Platform
8 select newer more appropriate JDK's for each platform. Fails for
unrecognized Platform versions so that we remember to select appropriate
JDKs.

Selecting JDK 11 for SLES 12 means that we will only be compatible with
the most recent (and supported) SP 5.