(PE-36861) Add a needrestart conf file for debian packages
In Ubuntu 22, needrestart is now called by default whenever a package is installed via apt. Because our packages lay down a temp directory for use by the service, when the services create and modify files in those temp dirs, needrestart thinks the service needs to be restarted. Because we manage restarting our services as needed, this drops a file in /etc/needrestart/conf.d that adds the service to the list of services needrestart should ignore.